[#48] refactor: ProductService의 내용을 ProductRetriever로 이동
